earthquake_cdc_pipeline/requirements.txt
requests==2.32.4
mysql-connector-python==9.3.0